solvebio/contrib/django_solvebio/test_requirements.txt
django-nose
coverage
django-coverage
django-jasmine
ipdb
flake8
fabric
mock
selenium
